+++ migrations/20250711180600_dao-vote-add-weight.up.sql
@@ -1,3 +1,3 @@
 -- Add up migration script here
 
-ALTER TABLE dao_vote ADD COLUMN IF NOT EXISTS weight BIGINT;
+ALTER TABLE dao_vote ADD COLUMN IF NOT EXISTS weight REAL;
